requirements.txt
openai==2.21.0
langchain==1.2.10
langchain-core==1.2.13
langchain-openai==1.1.10
langchain-community==0.4.1
faiss-cpu==1.13.2
pypdf==6.7.1
python-dotenv==1.0.1